Split value graph implementation modules
